% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{add_interference_indicators}
\alias{add_interference_indicators}
\title{Add indicators of infection on other leaves with a plant}
\usage{
add_interference_indicators(.data, .cutpoint = 0.25)
}
\arguments{
\item{.data}{\code{multistate_basis_pav.rds}}

\item{.cutpoint}{the threshold at which to define the indicators}
}
\description{
Add indicators of infection on other leaves with a plant
}
