% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{add_states}
\alias{add_states}
\title{Add indicators of infection on other leaves with a plant}
\usage{
add_states(.data, .states = c("C", "R"))
}
\arguments{
\item{.data}{\code{multistate_basis_pav.rds}}

\item{.states}{the names of variables to use in defining states. Note that death
is automatically included.}
}
\description{
Add indicators of infection on other leaves with a plant
}
