% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_msm_results.R
\name{compare_probs}
\alias{compare_probs}
\title{Extract probabilities from an msm model}
\usage{
compare_probs(msmfit, t, from, to, state_labels, .covariates = "mean")
}
\arguments{
\item{msmfit}{a msm}

\item{t}{vector of times}

\item{from}{from stqte}

\item{to}{to state}

\item{state_labels}{vector of labels for all states}

\item{.covariates}{a list of covariate levels or \code{"mean"}}
}
\description{
Extract probabilities from an msm model
}
