% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{create_q_dataset}
\alias{create_q_dataset}
\title{Create a data frame of possible transitions}
\usage{
create_q_dataset(qmatrix, states)
}
\arguments{
\item{qmatrix}{a Q matrix}

\item{states}{a data.frame of possible states}
}
\description{
Create a data frame of possible transitions
}
