% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_msm_models.R
\name{define_transitions}
\alias{define_transitions}
\alias{compare_reference}
\alias{find_noreturn_states}
\alias{create_possible_transition_matrix}
\title{Find other states that a reference state cannot go to}
\usage{
compare_reference(ref, others)

find_noreturn_states(ref_index, no_return_states, state_dt)

create_possible_transition_matrix(exclude_states_list)
}
\arguments{
\item{ref}{a vector of reference states}

\item{others}{a matrix of other states}
}
\description{
Find other states that a reference state cannot go to
}
