% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_msm_models.R
\name{fit_msm}
\alias{fit_msm}
\title{Fit an MSM for the PAV data}
\usage{
fit_msm(formula, subject, data, covariates = NULL, startQ = NULL,
  Q = NULL, censor = NULL, ...)
}
\description{
TODO: Fill in params
}
