% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{prepare_analysis_data}
\alias{prepare_analysis_data}
\title{Prepare analysis data objects}
\usage{
prepare_analysis_data(.data, states_of_interest)
}
\arguments{
\item{.data}{\code{multistate_basis_pav.rds}}

\item{states_of_interest}{the names of variables to use in defining states. Note that death
is automatically included.}
}
\description{
Prepare analysis data objects
}
