% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{view_q_indices}
\alias{view_q_indices}
\title{View the indices of the transitions}
\usage{
view_q_indices(qmatrix)
}
\arguments{
\item{qmatrix}{a Q matrix}
}
\description{
View the indices of the transitions
}
